Добавлен TravelPoint и TravelRoute объекты
