--- Assets/Prefabs/Guard/GoesToThatPosition.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1a31010399df7924ba43e024a3ee78f2
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
