+++ Assets/Prefabs/Guard/TravelPoint.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8aa509954bf44a24495611f972035caf
+guid: a1d3b17811662c54fab06386a2119acc
 PrefabImporter:
   externalObjects: {}
   userData:
