+++ Assets/Scripts/Guard.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 2fdfb1a05aeef624e940b9773c500e21
-PrefabImporter:
+guid: a4aa9bea49886254e815469119d20b92
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
