+++ Assets/Scripts/Guard/TravelPoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8649df8dca4122b4485805924d1edd12
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
