+++ Assets/Scripts/Guard/TravelRoute.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 23ed75e495dd09a4e80a88e3376f692d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
